Fixed wrong casing in component classes
from lowercase to TitleCase
