+++ version.md
@@ -1,1 +1,1 @@
-1.0.1
+1.0.2
